refactor: reuse what the helpdesk already ships
Three duplicates fold into existing pieces of the app:

- the portal's stand-in hi/fr dictionary becomes a fetch of
  helpdesk.api.general.get_translations — the same allow_guest,
  .po-backed map the agent portal reads, so every language the site
  ships works instead of two hand-typed ones. t()/setLanguage keep
  their contracts; a string the .po files lack falls back to English.
- the filter support's hand-ported useDebounceFn re-exports frappe-ui's
  debounce, which the Studio build already resolves (KbOrgTickets was
  importing it all along).
- get_agent_avatars moves from api/session.py to api/agent.py, beside
  the other agent reads it belongs with.
